changed file organisation
