+++ src/IO/ftiff.h
@@ -46,7 +46,6 @@
 #endif
 #include <assert.h>
 #include <string.h>
-#include "../utility.h"
 
 #define MBR_CNT(TYPE) int TYPE ## _mbrCnt
 #define MBR_BLK_LEN(TYPE) int TYPE ## _mbrBlkLen[]
