+++ src/interface.h
@@ -44,31 +44,35 @@ extern "C"
 #endif
 #include "fti.h"
 
+#include "fortran/ftif.h"
+
 #include "util/ini.h"
-#include "dataset.h"
-#include "keymap.h"
+#include "util/dataset.h"
+#include "util/keymap.h"
+#include "util/metaqueue.h"
+#include "util/macros.h"
+#include "util/utility.h"
+#include "util/failure-injection.h"
+
+#include "IO/posix.h"
+#include "IO/posix-dcp.h"
+#include "IO/hdf5-fti.h"
+#include "IO/ftiff.h"
+#include "IO/ftiff-dcp.h"
+
 #include "meta.h"
 #include "api-cuda.h"
 #include "postreco.h"
-#include "tools.h"
+#include "util/tools.h"
 #include "dcp.h"
 #include "conf.h"
 #include "checkpoint.h"
 #include "stage.h"
 #include "fti-io.h"
 #include "topo.h"
-#include "IO/posix.h"
-#include "IO/posix-dcp.h"
-#include "IO/hdf5-fti.h"
-#include "IO/ftiff.h"
-#include "IO/ftiff-dcp.h"
-#include "failure-injection.h"
 #include "postckpt.h"
 #include "recover.h"
-#include "fortran/ftif.h"
 #include "icp.h"
-#include "macros.h"
-#include "utility.h"
 
 #include "../deps/md5/md5.h"
 #include "../deps/iniparser/iniparser.h"
