+++ src/util/failure-injection.c
