+++ src/util/macros.c
