+++ src/util/tools.c
