+++ src/util/tools.h
