+++ src/util/utility.c
